+++ packages/react-server/src/logger/index.ts
@@ -1,8 +1,1 @@
-export {
-  createLogger,
-  Verbosity,
-  setLogger,
-  getLogger,
-  Logger,
-  noopLogger,
-} from './logger';
+export {requestLogger, Logger, getLogger} from './logger';