+++ packages/react-server/src/render/render.tsx
@@ -102,12 +102,15 @@ export function createRender(render: RenderFunction, options: Options = {}) {
       ctx.set(Header.ContentType, 'text/html');
       ctx.body = response;
     } catch (error) {
-      logger.error(error);
+      const errorMessage = `React server-side rendering error:\n${error.stack ||
+        error.message}`;
+
+      logger.log(errorMessage);
+      ctx.status = StatusCode.InternalServerError;
+
       // eslint-disable-next-line no-process-env
       if (process.env.NODE_ENV === 'development') {
-        ctx.body = `Internal Server Error: \n\n${error.message} \n\n ${
-          error.stack
-        }`;
+        ctx.body = errorMessage;
       } else {
         ctx.throw(StatusCode.InternalServerError, error);
       }