+++ packages/react-server/src/server/server.ts
@@ -7,7 +7,7 @@ import compose from 'koa-compose';
 import mount from 'koa-mount';
 import {middleware as sewingKitMiddleware} from '@shopify/sewing-kit-koa';
 import {createRender, RenderFunction} from '../render';
-import {createLogger, Verbosity} from '../logger';
+import {requestLogger} from '../logger';
 import {ping} from '../ping';
 
 const logger = console;
@@ -18,7 +18,6 @@ type Options = {
   assetPrefix?: string;
   serverMiddleware?: compose.Middleware<Context>[];
   render: RenderFunction;
-  debug?: boolean;
 };
 
 /**
@@ -27,13 +26,13 @@ type Options = {
  * @returns a Server instance
  */
 export function createServer(options: Options): Server {
-  const {port, assetPrefix, render, debug, serverMiddleware, ip} = options;
+  const {port, assetPrefix, render, serverMiddleware, ip} = options;
   const app = new Koa();
   const manifestPath = getManifestPath(process.cwd());
 
   app.use(mount('/services/ping', ping));
 
-  app.use(createLogger({level: debug ? Verbosity.Debug : Verbosity.Off}));
+  app.use(requestLogger);
   app.use(sewingKitMiddleware({assetPrefix, manifestPath}));
 
   if (serverMiddleware) {