+++ packages/react-server/CHANGELOG.md
@@ -7,6 +7,10 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+## [0.5.1] - 2019-09-11
+
+- Add spacing between "[React Server]" prefix and logs [#984](https://github.com/Shopify/quilt/pull/984)
+
 ## [0.5.0] - 2019-09-11
 
 ### Added